requirements.txt
# Django and REST Framework
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

# Authentication
PyJWT==2.8.0

# Database and utilities
sqlparse==0.5.0
tzdata==2024.1
python-decouple==3.8

# Image handling - Use version 11+ for Python 3.13 compatibility
Pillow>=11.0.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Testing
pytest==8.0.0
pytest-django==4.8.0